+++ CHANGELOG.md
@@ -5,6 +5,7 @@ _WIP_
 
 **New features**
 - Added flags for explicitly setting dark or light color scheme
+- Number of replies now also indicate how many of the replies are new since last visit
 
 ## 2.1
 _25.06.22_